requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
python-dotenv==1.0.0
websockets==12.0
httpx==0.27.2 
faiss-cpu
sentence-transformers
numpy
pydantic
torch